+++ src/gaViz/main/Population.java
@@ -84,7 +84,7 @@ public class Population {
 	}
 	
 	public Individual getProbabilisticIndividual () {
-		float randVal = (float) Math.random();
+		double randVal = Math.random();
 		for (int i = 0; i < size; i++) {
 			if (randVal >= this.individuals[i].getCumulativeLowerBound() && 
 				randVal <= this.individuals[i].getCumulativeUpperBound() ) {
